scripts/utils/check_imports.py
"""
Module description

Purpose: analyzed as source_code
Expected Lifetime: permanent
Category: SOURCE_CODE
Created: 2026-01-21
Last Updated: 2026-01-24
"""
def main():
    try:
        import streamlit
        print("streamlit: OK")
    except ImportError as e:
        print(f"streamlit: FAIL {e}")

    try:
        import streamlit_authenticator
        print("streamlit_authenticator: OK")
    except ImportError as e:
        print(f"streamlit_authenticator: FAIL {e}")

    try:
        import yaml
        print("yaml: OK")
    except ImportError as e:
        print(f"yaml: FAIL {e}")

if __name__ == "__main__":
    main()
