requirements.txt
# CV Generator Dependencies
# Install with: pip install -r requirements.txt

# Core dependencies
python-docx>=1.2.0

# OpenAI API for PDF to JSON conversion
openai>=2.14.0

# PDF processing
pypdf>=6.5.0

# Environment variables
python-dotenv>=1.2.1

# Image processing (for logo optimization)
Pillow>=12.0.0

# Excel file generation
openpyxl>=3.10.0

# Web Interface
streamlit>=1.52.2

# Authentication & Security
streamlit-authenticator>=0.4.2
PyYAML>=6.0.3
bcrypt>=5.0.0
extra-streamlit-components>=0.1.81

# Auto-added dependencies
altair>=6.0.0
annotated-types>=0.7.0
anyio>=4.12.0
attrs>=25.4.0
blinker>=1.9.0
cachetools>=6.2.4
captcha>=0.7.1
certifi>=2025.11.12
cffi>=2.0.0
charset-normalizer>=3.4.4
click>=8.3.1
colorama>=0.4.6
coverage>=7.13.0
cryptography>=46.0.3
distro>=1.9.0
gitdb>=4.0.12
gitpython>=3.1.45
h11>=0.16.0
httpcore>=1.0.9
httpx>=0.28.1
idna>=3.11
iniconfig>=2.3.0
jinja2>=3.1.6
jiter>=0.12.0
jsonschema>=4.25.1
jsonschema-specifications>=2025.9.1
lxml>=6.0.2
markupsafe>=3.0.3
mccabe>=0.7.0
mypy_extensions>=1.1.0
narwhals>=2.14.0
numpy>=2.4.0
packaging>=25.0
pandas>=2.3.3
pathspec>=0.12.1
platformdirs>=4.5.1
plotly>=6.5.0
pluggy>=1.6.0
protobuf>=6.33.2
pyarrow>=22.0.0
pycodestyle>=2.11.1
pycparser>=2.23
pydantic>=2.12.5
pydantic_core>=2.41.5
pydeck>=0.9.1
pyflakes>=3.2.0
pygments>=2.19.2
pyjwt>=2.10.1
python-dateutil>=2.9.0.post0
pytz>=2025.2
referencing>=0.37.0
requests>=2.32.5
rpds-py>=0.30.0
six>=1.17.0
smmap>=5.0.2
sniffio>=1.3.1
tenacity>=9.1.2
toml>=0.10.2
tornado>=6.5.4
tqdm>=4.67.1
typing-inspection>=0.4.2
typing_extensions>=4.15.0
tzdata>=2025.3
urllib3>=2.6.2
watchdog>=6.0.0
